:construction: Dockerize, fixes DP-417
Dockerize the project via `docker-compose.yml`.
The compose file consists in two containers:
- frontend: serves the pages with PHP and consumes JSON files
- worker: pulls backend data from Redis and write to JSON files

Both containers need to be in the same network as the backend containers.
The frontend to use the shared MariaDB database and the worker to use
the shared Redis cache.

Also make it possible to dynamically change `JSON_LOCATION` and `REDIS_HOST`
via environment variables so manually editing is no longer needed.